Refactor openai messages conversion
